[NA] : better operator

+++ src/memory/Memory.cpp
@@ -105,7 +105,7 @@ namespace AbstractVM
     {
         int maxPrecs = 0;
         bool pre = false;
-        int lastZero = 0;
+        size_t lastZero = 0;
         bool isLastZero = false;
         for (int i = ope->toString().size()-1; i >= 0; i--) {
             if (ope->toString().at(i) == '\0' || ope->toString().at(i) == '0' || ope->toString().at(i) == '.') {
